pyproject.toml
[tool.poetry]
name = "saleor-export"
version = "0.1.0"
description = "An app that allows sellers exporting certain data to a file."
authors = ["Anatolii Serebrianskii <anatolii.serebrianskii@tango.agency>"]

[tool.poetry.dependencies]
python = "^3.9"
celery = "^5.1"
sqlmodel = "^0.0.6"
strawberry-graphql = "^0.93.21"
saleor-app-base = {git = "git@github.com:SaleorIntegrations/saleor-app-base.git", rev = "main"}
greenlet = "^1.1.2"
asgiref = "^3.4.1"
redis = "^4.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
pytest-asyncio = "^0.16.0"
pytest-mock = "^3.6.1"
pre-commit = "^2.16.0"
coverage = "^6.2"
httpx = "0.18.1"
"testing.postgresql" = "^1.3.0"
asgiref = "^3.4.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
